Add basic Docker setup
This adds NGINX to act as a proxy to send requests towards the `api`
container.

The `api` image is build using two stages. The first stage is used to
build the Rust binary (first dependencies, then the actual binary).

Finally, the Redis image constains the Redis server with the RedisBloom
module. It contains a Redis dump file that contains the `pwned-bloom`
bloom filter.