Add form to allow people to check passwords
Adds a simple page to `/` (directly served by NGINX) that allows people
to submit a password to be checked against the API.

The password is hashes locally using the `SubtleCrypto` module. If the
hash is present in the bloom filter, the user will be advised to not
use this password and change their passwords if they use this password.

If the password does not appear in the bloom filter, this will be
communicated with the user. However, we also note that their password
may actually not be secure.

In both cases we recommend the use of a password manager to create
(secure) random passwords.