Re-bloom for June 2024 dump
We have a new dump from Pwned Passwords dated 2025-06-05 containing
`936_491_757` SHA1 hashes. That is `89_268_355` more pwned
passwords than in March 2022. As such it is time to create a new
bloom filter for the API.

The parameters for the re-bloom are as follows:

```txt
m = 35905312644 (4.18GiB)
n = 936491757
k = 27
p = 0.00000001 (1 in 99'857'412)
```

In other words, we keep the same number of hashing functions and
false-positive probability but increase our RAM usage (by 0.4GiB).

The process for re-blooming and the new Rust script for this are
documented in the `README` of the `data` subdirectory.

---

Also updates all dependencies and moves the images to Alpine 3.20.